[FEAT] wrapped with suspense n utils into ts
