+++ src/app/(user)/infodesk/notices/[id]/GetUserNoticeDetail.ts
